A couple updates
Move the S3 output to just the root, without /api/where to sync with
the proposal changes.
Also, handle an empty bounds string. I'm not sure if this is the right
thing to do, or whether the script should automatically mark it as
inactive. But for now we'll attempt to propogate the information even
if we think it's invalid.